+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Previous classification is not required if changes are simple or all belong to t
 - New text prompt function for translate texts, `Translate`.
 - Added an example of using `Translate` in `Encamina.Enmarcha.Samples.SemanticKernel.Text`.
 - Bug fix: Temporary workaround for handling Http NotFound exception in `MemoryStoreExtender`. [(#72)](https://github.com/Encamina/enmarcha/issues/72)
+- Added new method `ExistsMemoryAsync` in `MemoryStoreExtender`.
 
 ## [8.1.2]